# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/masjids": {
      "get": {
        "operationId": "MasjidsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all masjids",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MasjidResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all masjids",
        "tags": [
          "masjids"
        ]
      },
      "post": {
        "operationId": "MasjidsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMasjidDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The masjid has been created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasjidResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new masjid",
        "tags": [
          "masjids"
        ]
      }
    },
    "/masjids/{id}": {
      "get": {
        "operationId": "MasjidsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Masjid ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The masjid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasjidResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Masjid not found"
          }
        },
        "summary": "Get a masjid by ID",
        "tags": [
          "masjids"
        ]
      },
      "patch": {
        "operationId": "MasjidsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Masjid ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMasjidDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The masjid has been updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasjidResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden resource"
          },
          "404": {
            "description": "Masjid not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a masjid",
        "tags": [
          "masjids"
        ]
      },
      "delete": {
        "operationId": "MasjidsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Masjid ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The masjid has been deleted"
          },
          "403": {
            "description": "Forbidden resource"
          },
          "404": {
            "description": "Masjid not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a masjid",
        "tags": [
          "masjids"
        ]
      }
    },
    "/masjids/{masjidId}/admins": {
      "get": {
        "operationId": "MasjidsController_findAllAdmins",
        "parameters": [
          {
            "name": "masjidId",
            "required": true,
            "in": "path",
            "description": "Masjid ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of masjid admins",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MasjidAdminResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Masjid not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all admins for a masjid",
        "tags": [
          "masjids"
        ]
      },
      "post": {
        "operationId": "MasjidsController_addAdmin",
        "parameters": [
          {
            "name": "masjidId",
            "required": true,
            "in": "path",
            "description": "Masjid ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMasjidAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The admin has been added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasjidAdminResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Masjid or user not found"
          },
          "409": {
            "description": "User is already an admin"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an admin to a masjid",
        "tags": [
          "masjids"
        ]
      }
    },
    "/masjids/{masjidId}/admins/{adminId}": {
      "get": {
        "operationId": "MasjidsController_findOneAdmin",
        "parameters": [
          {
            "name": "masjidId",
            "required": true,
            "in": "path",
            "description": "Masjid ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "description": "Admin ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The masjid admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasjidAdminResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Admin not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific admin for a masjid",
        "tags": [
          "masjids"
        ]
      },
      "patch": {
        "operationId": "MasjidsController_updateAdmin",
        "parameters": [
          {
            "name": "masjidId",
            "required": true,
            "in": "path",
            "description": "Masjid ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "description": "Admin ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMasjidAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The admin's role has been updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasjidAdminResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Admin not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an admin's role",
        "tags": [
          "masjids"
        ]
      },
      "delete": {
        "operationId": "MasjidsController_removeAdmin",
        "parameters": [
          {
            "name": "masjidId",
            "required": true,
            "in": "path",
            "description": "Masjid ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "description": "Admin ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The admin has been removed"
          },
          "403": {
            "description": "Cannot remove the last admin"
          },
          "404": {
            "description": "Admin not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove an admin from a masjid",
        "tags": [
          "masjids"
        ]
      }
    },
...
    "/campaigns/masjid/{masjidId}": {
      "get": {
        "operationId": "CampaignsController_findByMasjid",
        "parameters": [
          {
            "name": "masjidId",
            "required": true,
            "in": "path",
            "description": "Masjid ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of campaigns for the masjid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get campaigns by masjid",
        "tags": [
          "campaigns"
        ]
      }
    },
...
  },
  "info": {
    "title": "Masjid Network API",
    "description": "\n# Masjid Network API\n\n## Overview\nThe Masjid Network API provides a comprehensive backend for managing mosques (masjids), fundraising campaigns, donations, and user accounts. This platform connects mosques with their communities and facilitates fundraising for various projects and initiatives.\n\n## Features\n- User authentication and management\n- Masjid/mosque profiles and administration\n- Fundraising campaigns\n- Donation processing\n- Payment handling\n\n## Authentication\nMost endpoints require authentication using JWT Bearer tokens.\n\nTo obtain a token, use the /auth/login endpoint with valid credentials.\nInclude the token in the Authorization header: `Bearer your_token_here`\n\n## Getting Started\n1. Register a user account via /auth/register\n2. Login to obtain a JWT token via /auth/login\n3. Use the token to access protected endpoints\n    ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "MasjidResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "example": "Al-Aqsa Masjid"
          },
          "description": {
            "type": "object",
            "example": "A beautiful masjid serving the local community"
          },
          "address": {
            "type": "string",
            "example": "123 Main St"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "example": "NY"
          },
          "country": {
            "type": "string",
            "example": "USA"
          },
          "zipCode": {
            "type": "string",
            "example": "10001"
          },
          "email": {
            "type": "object",
            "example": "info@alaqsamasjid.org"
          },
          "phone": {
            "type": "object",
            "example": "+1234567890"
          },
          "website": {
            "type": "object",
            "example": "https://alaqsamasjid.org"
          },
          "logoUrl": {
            "type": "object",
            "example": "https://example.com/logo.jpg"
          },
          "coverImageUrl": {
            "type": "object",
            "example": "https://example.com/cover.jpg"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "isVerified": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "city",
          "state",
          "country",
          "zipCode",
          "createdAt",
          "updatedAt",
          "isVerified"
        ]
      },
      "CreateMasjidDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Al-Aqsa Masjid"
          },
          "description": {
            "type": "string",
            "example": "A beautiful masjid serving the local community"
          },
          "address": {
            "type": "string",
            "example": "123 Main St"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "example": "NY"
          },
          "country": {
            "type": "string",
            "example": "USA"
          },
          "zipCode": {
            "type": "string",
            "example": "10001"
          },
          "email": {
            "type": "string",
            "example": "info@alaqsamasjid.org"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "website": {
            "type": "string",
            "example": "https://alaqsamasjid.org"
          },
          "logoUrl": {
            "type": "string",
            "example": "https://example.com/logo.jpg"
          },
          "coverImageUrl": {
            "type": "string",
            "example": "https://example.com/cover.jpg"
          }
        },
        "required": [
          "name",
          "address",
          "city",
          "state",
          "country",
          "zipCode"
        ]
      },
      "UpdateMasjidDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Al-Aqsa Masjid"
          },
          "description": {
            "type": "string",
            "example": "A beautiful masjid serving the local community"
          },
          "address": {
            "type": "string",
            "example": "123 Main St"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "example": "NY"
          },
          "country": {
            "type": "string",
            "example": "USA"
          },
          "zipCode": {
            "type": "string",
            "example": "10001"
          },
          "email": {
            "type": "string",
            "example": "info@alaqsamasjid.org"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "website": {
            "type": "string",
            "example": "https://alaqsamasjid.org"
          },
          "logoUrl": {
            "type": "string",
            "example": "https://example.com/logo.jpg"
          },
          "coverImageUrl": {
            "type": "string",
            "example": "https://example.com/cover.jpg"
          },
          "isVerified": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "MasjidAdminResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "role": {
            "type": "string",
            "example": "ADMIN"
          },
          "userId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "masjidId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "role",
          "userId",
          "masjidId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateMasjidAdminDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "role": {
            "type": "string",
            "example": "ADMIN"
          }
        },
        "required": [
          "userId"
        ]
      },
      "UpdateMasjidAdminDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "example": "ADMIN"
          }
        }
      },
      "CampaignResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the campaign"
          },
          "title": {
            "type": "string",
            "description": "Title of the campaign"
          },
          "description": {
            "type": "string",
            "description": "Description of the campaign"
          },
          "goal": {
            "type": "number",
            "description": "Funding goal amount for the campaign"
          },
          "raised": {
            "type": "number",
            "description": "Current amount raised for the campaign"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "description": "Start date of the campaign"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "End date of the campaign (optional)"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the campaign is active"
          },
          "coverImageUrl": {
            "type": "string",
            "description": "Cover image URL for the campaign"
          },
          "masjidId": {
            "type": "string",
            "description": "ID of the masjid running the campaign"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the campaign was created"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the campaign was last updated"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "goal",
          "raised",
          "startDate",
          "endDate",
          "isActive",
          "coverImageUrl",
          "masjidId",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}